Convert uints to size_ts
Review URL: http://codereview.chromium.org/115429

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@16209 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/browser_theme_provider.cc
@@ -94,7 +94,7 @@ BrowserThemeProvider::BrowserThemeProvider()
     : rb_(ResourceBundle::GetSharedInstance()) {
   static bool initialized = false;
   if (!initialized) {
-    for (unsigned int i = 0; i < sizeof(kToolbarButtonIDs); ++i) {
+    for (size_t i = 0; i < sizeof(kToolbarButtonIDs); ++i) {
       button_images_[kToolbarButtonIDs[i]] = true;
     }
     frame_tints_[IDR_THEME_FRAME] = TINT_FRAME;
